Update battleship.py
Refactored with the structure that a player has offense and defense strategies. Strategies make decisions about where to place ships (defense) and where to shoot (offense), but do not change the board. The player object changes the board by taking shots.

Player objects must now be instances of HumanPlayer or AIPlayer, which are subclasses of the abstract base class Player.
There are now also Offense and Defense abstract base classes, which can be subclassed to implement specific strategies.

Added a Coord class that is a container for a location on board of a particular size. This allows for easy translation between row/column format like (2,5) and letter/number format like C6.

This version is a work in progress. Offenses will not necessarily work as of the current commit.